Add option to use existing VPC
AWS resources will be put it an existing VPC if all parameters named "Existing*"
are given.
We want to reuse a VPC when e.g. we need a static IP inside Metaflow steps
(that's why AWS::Batch::ComputeEnvironment can be put inside a private subnet).
With a static, and whitelisted, IP we can access protected resource during a
Metaflow run.

The changes are specific to the VPC requirements in the Laika project.